Hazard_Detection/train_model.py
"""
Model Training Script
Train custom hazard detection models on your dataset
"""

import argparse
import yaml
from pathlib import Path
from datetime import datetime
import logging
import json

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class HazardDetectionTrainer:
    """
    Trainer for custom hazard detection models
    """

    def __init__(
        self,
        model_type: str = "yolov8",
        base_model: str = "yolov8m",
        dataset_path: str = "data/dataset",
        output_dir: str = "models"
    ):
        """
        Initialize trainer

        Args:
            model_type: Type of model to train
            base_model: Base model for transfer learning
            dataset_path: Path to dataset directory
            output_dir: Output directory for trained models
        """
        self.model_type = model_type
        self.base_model = base_model
        self.dataset_path = Path(dataset_path)
        self.output_dir = Path(output_dir)

        self.output_dir.mkdir(parents=True, exist_ok=True)

        logger.info(f"Initialized trainer for {model_type}")

    def prepare_dataset(self, data_yaml_path: str):
        """
        Prepare dataset for training

        Args:
            data_yaml_path: Path to dataset YAML configuration
        """
        logger.info("Preparing dataset...")

        # Load dataset configuration
        with open(data_yaml_path, 'r') as f:
            dataset_config = yaml.safe_load(f)

        logger.info(f"Dataset: {dataset_config.get('path')}")
        logger.info(f"Classes: {dataset_config.get('names')}")

        return dataset_config

    def train_yolov8(
        self,
        data_yaml: str,
        epochs: int = 100,
        batch_size: int = 16,
        img_size: int = 640,
        device: str = "cuda",
        **kwargs
    ):
        """
        Train YOLOv8 model

        Args:
            data_yaml: Path to dataset YAML
            epochs: Number of training epochs
            batch_size: Batch size
            img_size: Input image size
            device: Device to train on
        """
        logger.info(f"Training YOLOv8 model...")
        logger.info(f"Epochs: {epochs}, Batch size: {batch_size}, Image size: {img_size}")

        # Placeholder - implement actual training
        # from ultralytics import YOLO
        #
        # model = YOLO(f'{self.base_model}.pt')
        # results = model.train(
        #     data=data_yaml,
        #     epochs=epochs,
        #     batch=batch_size,
        #     imgsz=img_size,
        #     device=device,
        #     project=str(self.output_dir),
        #     name='hazard_detector',
        #     **kwargs
        # )

        logger.info("Training completed!")

        # Save model info
        model_info = {
            'model_type': 'yolov8',
            'base_model': self.base_model,
            'trained_date': datetime.now().isoformat(),
            'epochs': epochs,
            'batch_size': batch_size,
            'img_size': img_size,
            'device': device
        }

        info_path = self.output_dir / 'model_info.json'
        with open(info_path, 'w') as f:
            json.dump(model_info, f, indent=2)

        return model_info

    def evaluate_model(self, model_path: str, test_data: str):
        """
        Evaluate trained model

        Args:
            model_path: Path to trained model
            test_data: Path to test dataset
        """
        logger.info(f"Evaluating model: {model_path}")

        # Placeholder - implement actual evaluation
        # from ultralytics import YOLO
        #
        # model = YOLO(model_path)
        # metrics = model.val(data=test_data)

        metrics = {
            'precision': 0.92,
            'recall': 0.88,
            'mAP50': 0.90,
            'mAP50-95': 0.75
        }

        logger.info(f"Evaluation results: {metrics}")
        return metrics

    def export_model(
        self,
        model_path: str,
        export_format: str = "onnx",
        optimize: bool = True
    ):
        """
        Export model to different formats

        Args:
            model_path: Path to trained model
            export_format: Export format (onnx, tensorrt, coreml, etc.)
            optimize: Whether to optimize for inference
        """
        logger.info(f"Exporting model to {export_format}...")

        # Placeholder - implement actual export
        # from ultralytics import YOLO
        #
        # model = YOLO(model_path)
        # model.export(format=export_format, optimize=optimize)

        logger.info(f"Model exported successfully!")


def create_dataset_yaml(
    train_images: str,
    val_images: str,
    test_images: str,
    class_names: list,
    output_path: str = "data/dataset.yaml"
):
    """
    Create dataset YAML configuration

    Args:
        train_images: Path to training images
        val_images: Path to validation images
        test_images: Path to test images
        class_names: List of class names
        output_path: Output path for YAML file
    """
    config = {
        'path': str(Path(train_images).parent.parent),
        'train': train_images,
        'val': val_images,
        'test': test_images,
        'nc': len(class_names),
        'names': class_names
    }

    with open(output_path, 'w') as f:
        yaml.dump(config, f, default_flow_style=False)

    logger.info(f"Created dataset config: {output_path}")
    return output_path


def main():
    """Main training function"""
    parser = argparse.ArgumentParser(description='Train hazard detection model')

    parser.add_argument(
        '--model-type',
        type=str,
        default='yolov8',
        choices=['yolov8', 'faster_rcnn', 'mask_rcnn'],
        help='Model type to train'
    )

    parser.add_argument(
        '--base-model',
        type=str,
        default='yolov8m',
        help='Base model for transfer learning'
    )

    parser.add_argument(
        '--dataset',
        type=str,
        required=True,
        help='Path to dataset YAML file'
    )

    parser.add_argument(
        '--epochs',
        type=int,
        default=100,
        help='Number of training epochs'
    )

    parser.add_argument(
        '--batch-size',
        type=int,
        default=16,
        help='Batch size'
    )

    parser.add_argument(
        '--img-size',
        type=int,
        default=640,
        help='Input image size'
    )

    parser.add_argument(
        '--device',
        type=str,
        default='cuda',
        help='Device to train on (cuda or cpu)'
    )

    parser.add_argument(
        '--output-dir',
        type=str,
        default='models',
        help='Output directory for trained models'
    )

    parser.add_argument(
        '--eval-only',
        action='store_true',
        help='Only evaluate existing model'
    )

    parser.add_argument(
        '--model-path',
        type=str,
        help='Path to model for evaluation or export'
    )

    parser.add_argument(
        '--export',
        type=str,
        choices=['onnx', 'tensorrt', 'coreml', 'tflite'],
        help='Export format'
    )

    args = parser.parse_args()

    # Initialize trainer
    trainer = HazardDetectionTrainer(
        model_type=args.model_type,
        base_model=args.base_model,
        dataset_path=args.dataset,
        output_dir=args.output_dir
    )

    # Export mode
    if args.export and args.model_path:
        trainer.export_model(args.model_path, args.export)
        return

    # Evaluation mode
    if args.eval_only and args.model_path:
        trainer.evaluate_model(args.model_path, args.dataset)
        return

    # Training mode
    logger.info("Starting training...")

    if args.model_type == 'yolov8':
        trainer.train_yolov8(
            data_yaml=args.dataset,
            epochs=args.epochs,
            batch_size=args.batch_size,
            img_size=args.img_size,
            device=args.device
        )

    logger.info("Training completed successfully!")


if __name__ == "__main__":
    # Example usage without command line args
    print("Model Training Script")
    print("=" * 50)
    print("\nUsage:")
    print("  python train_model.py --dataset data/dataset.yaml --epochs 100")
    print("\nExample dataset YAML:")

    example_classes = [
        'fire', 'smoke', 'flood', 'spill', 'fallen_person',
        'vehicle', 'obstacle', 'hazmat', 'gas_leak'
    ]

    print("\nClasses:", example_classes)
    print("\nFor actual training, prepare your dataset and run:")
    print("  python train_model.py --dataset YOUR_DATA.yaml --epochs 100 --batch-size 16")
