Hazard_Detection/requirements.txt
# Core Dependencies
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8
tensorflow>=2.13.0
onnxruntime>=1.15.0

# Computer Vision
opencv-contrib-python>=4.8.0
scikit-image>=0.21.0
albumentations>=1.3.0

# LLM Integration
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# API & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Data Processing
pandas>=2.0.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL
pymongo>=4.5.0  # MongoDB
redis>=5.0.0

# Notifications & Alerts
requests>=2.31.0
python-telegram-bot>=20.5
slack-sdk>=3.23.0
twilio>=8.9.0
sendgrid>=6.10.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.66.0
loguru>=0.7.0

# Tracking
filterpy>=1.4.5  # Kalman filtering for tracking

# Image Processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Async
aiohttp>=3.9.0
aiofiles>=23.2.0

# Monitoring
prometheus-client>=0.18.0
sentry-sdk>=1.32.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For testing FastAPI

# Development
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# ML Ops
wandb>=0.16.0  # Experiment tracking
mlflow>=2.8.0

# Optional GPU acceleration
# tensorrt>=8.6.0  # Uncomment for TensorRT support
# onnx>=1.14.0
